Split routing into modules.
